# docs/swagger.yaml
definitions:
  models.CreateRoomReqBody:
    properties:
      host_name:
        type: string
      max_peers:
        type: integer
      room_name:
        type: string
    type: object
  models.GetRoomsListRespBody:
    properties:
      data:
        items:
          $ref: '#/definitions/models.VoiceRoom'
        type: array
      message:
        type: string
    type: object
  models.MessageJson:
    properties:
      message:
        type: string
    type: object
  models.Peer:
    properties:
      descriptor:
        type: string
      id:
        type: string
      is_host:
        type: boolean
      name:
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
  models.VoiceRoom:
    properties:
      host_name:
        type: string
      id:
        type: string
      max_peers:
        type: integer
      name:
        type: string
      peers:
        items:
          $ref: '#/definitions/models.Peer'
        type: array
    type: object
info:
  contact: {}
paths:
  /voicechat/create:
    post:
      consumes:
      - application/json
      description: create new room
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateRoomReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.MessageJson'
      summary: Create room handler
      tags:
      - voicechat
  /voicechat/rooms:
    get:
      consumes:
      - application/json
      description: get rooms list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetRoomsListRespBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Rooms list handler
      tags:
      - voicechat
  /voicechat/ws/connect:
    get:
      consumes:
      - application/json
      description: connect to existing room
      parameters:
      - description: id of room
        in: query
        name: room_id
        required: true
        type: string
      - description: username
        in: query
        name: peer_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.MessageJson'
      summary: Connect to room handler
      tags:
      - voicechat
  /voicechat/ws/rooms:
    get:
      consumes:
      - application/json
      description: connect to socket to get rooms changes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.MessageJson'
      summary: Rooms list ws handler
      tags:
      - voicechat
swagger: "2.0"